Align with the new project structure
