+++ consumer/counter-consumer/src/main/java/io/pivotal/java/function/counter/consumer/StringToSpelConversionFunction.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2020 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.pivotal.java.function.counter.consumer;
+
+import java.util.function.Function;
+
+import org.springframework.expression.EvaluationContext;
+import org.springframework.expression.Expression;
+import org.springframework.expression.ParseException;
+import org.springframework.expression.spel.standard.SpelExpression;
+import org.springframework.expression.spel.standard.SpelExpressionParser;
+
+/**
+ * Converter from String to Spring Expression.
+ *
+ * TODO: This could be a top level project.
+ */
+public class StringToSpelConversionFunction implements Function<String, Expression> {
+
+	private final SpelExpressionParser parser;
+
+	private final EvaluationContext evaluationContext;
+
+	public StringToSpelConversionFunction(EvaluationContext evaluationContext) {
+		this(new SpelExpressionParser(), evaluationContext);
+	}
+
+	public StringToSpelConversionFunction(SpelExpressionParser parser, EvaluationContext evaluationContext) {
+		this.evaluationContext = evaluationContext;
+		this.parser = parser;
+	}
+
+	@Override
+	public Expression apply(String source) {
+		try {
+			Expression expression = parser.parseExpression(source);
+			if (expression instanceof SpelExpression) {
+				((SpelExpression) expression).setEvaluationContext(evaluationContext);
+			}
+			return expression;
+		}
+		catch (ParseException e) {
+			throw new IllegalArgumentException(String.format(
+					"Could not convert '%s' into a SpEL expression", source), e);
+		}
+	}
+}
