--- consumer/pom.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>java-functions-consumer</artifactId>
-	<version>1.0.0.BUILD-SNAPSHOT</version>
-	<name>java-functions-consumer</name>
-	<description>Pivotal Java Functions Consumers</description>
-	<packaging>pom</packaging>
-
-	<parent>
-		<groupId>io.pivotal.java.function</groupId>
-		<artifactId>java-functions-parent</artifactId>
-		<version>1.0.0.BUILD-SNAPSHOT</version>
-	</parent>
-
-	<properties>
-		<java.version>1.8</java.version>
-	</properties>
-
-  	<modules>
-		<module>cassandra-consumer</module>
-		<module>jdbc-consumer</module>
-		<module>log-consumer</module>
-		<module>mongodb-consumer</module>
-		<module>rabbit-consumer</module>
-		<module>counter-consumer</module>
-	</modules>
-
-</project>
