+++ pom.xml
@@ -21,6 +21,7 @@
 
   	<modules>
 		<module>consumer/cassandra-consumer</module>
+		<module>consumer/counter-consumer</module>
 		<module>consumer/jdbc-consumer</module>
 		<module>consumer/log-consumer</module>
 		<module>consumer/mongodb-consumer</module>
