--- supplier/pom.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-	<artifactId>java-functions-supplier</artifactId>
-	<version>1.0.0.BUILD-SNAPSHOT</version>
-	<name>java-functions-supplier</name>
-	<description>Pivotal Java Functions Suppliers</description>
-	<packaging>pom</packaging>
-
-	<parent>
-		<groupId>io.pivotal.java.function</groupId>
-		<artifactId>java-functions-parent</artifactId>
-		<version>1.0.0.BUILD-SNAPSHOT</version>
-	</parent>
-
-	<properties>
-		<java.version>1.8</java.version>
-		<spring-cloud-function.version>3.0.2.RELEASE</spring-cloud-function.version>
-	</properties>
-
-  	<modules>
-		<module>filter-function</module>
-		<module>spel-function</module>
-		<module>splitter-function</module>
-	</modules>
-
-</project>
